Changes: Acquisition contract moved in contract folder
