+++ app/src/Form/Type/Work/WorkFormType.php
@@ -6,7 +6,6 @@ namespace App\Form\Type\Work;
 
 use App\Entity\Work\Work;
 use App\Form\Dto\Work\WorkFormDto;
-use App\Form\Type\Common\BroadcastChannelAutocompleteField;
 use App\Form\Validator\Constraint\UniqueEntityField;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CountryType;
@@ -70,9 +69,6 @@ class WorkFormType extends AbstractType
             ])
             ->add('duration', TextType::class, [
                 'required' => false,
-            ])
-            ->add('broadcastChannels', BroadcastChannelAutocompleteField::class, [
-                'multiple' => true,
             ])
         ;
     }
